pyproject.toml
[project]
name = "nv-um"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.3",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi>=0.116.0",
    "fastapi-mail>=1.5.0",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
